Test class for attribute scope. Separated tests for entity scope into separate class.
